@@ -19,79 +18,61 @@
   ],
   "dependencies": {
     "@babel/polyfill": "^7.4.3",
-    "@eluvio/elv-client-js": "^4.0.86",
+    "@eluvio/elv-client-js": "^4.0.88",
+    "@mantine/core": "^6.0.21",
+    "@mantine/hooks": "^6.0.21",
+    "@tabler/icons-react": "^2.39.0",
     "buffer": "^5.6.0",
     "elv-components-js": "git+https://github.com/eluv-io/elv-components-js.git#686933243cda82288bb537a1bf1aba3c1c0263b3",
-    "mobx": "^5.13.0",
-    "mobx-react": "^6.1.3",
+    "mobx": "^6.10.2",
+    "mobx-react": "^9.0.1",
     "path": "^0.12.7",
     "prop-types": "^15.7.2",
     "react": "^16.8.6",
     "react-dom": "^16.8.6",
     "react-markdown": "^8.0.6",
-    "react-router": "^5.1.2",
-    "react-router-dom": "^5.1.2",
-    "url-join": "^4.0.0"
+    "react-router": "^6.16.0",
+    "react-router-dom": "^6.16.0",
+    "url-join": "^4.0.0",
+    "uuid": "^9.0.1"
   },
   "devDependencies": {
-    "@babel/core": "^7.4.3",
-    "@babel/plugin-proposal-object-rest-spread": "^7.4.3",
-    "@babel/plugin-syntax-dynamic-import": "^7.2.0",
-    "@babel/plugin-syntax-jsx": "^7.2.0",
-    "@babel/plugin-syntax-object-rest-spread": "^7.2.0",
-    "@babel/plugin-transform-react-jsx": "^7.3.0",
-    "@babel/plugin-transform-regenerator": "^7.4.3",
-    "@babel/plugin-transform-runtime": "^7.4.3",
-    "@babel/preset-env": "^7.20.2",
+    "@babel/core": "^7.23.0",
+    "@babel/preset-env": "^7.22.20",
     "@babel/preset-react": "^7.0.0",
-    "@babel/runtime": "^7.4.3",
-    "autoprefixer": "^9.5.0",
+    "@babel/runtime": "^7.23.1",
+    "@pmmmwh/react-refresh-webpack-plugin": "^0.5.11",
+    "autoprefixer": "^10.4.16",
     "babel-cli": "^6.26.0",
     "babel-core": "^7.0.0-bridge.0",
     "babel-eslint": "^10.0.1",
-    "babel-loader": "^8.0.5",
+    "babel-loader": "^9.1.3",
     "babel-plugin-syntax-async-functions": "^6.13.0",
-    "babel-preset-mobx": "^2.0.0",
     "copy-webpack-plugin": "^4.6.0",
-    "css-loader": "^1.0.1",
+    "css-loader": "^6.8.1",
     "csv-parse": "^4.16.3",
     "eslint": "^5.16.0",
     "eslint-plugin-react": "^7.12.4",
-    "file-loader": "^2.0.0",
     "html-webpack-inline-source-plugin": "0.0.10",
-    "html-webpack-plugin": "^3.2.0",
-    "postcss-loader": "^3.0.0",
-    "pre-commit": "^1.2.2",
+    "html-webpack-plugin": "^5.5.3",
+    "postcss-loader": "^7.3.3",
     "raw-loader": "^0.5.1",
+    "react-refresh": "^0.14.0",
     "regenerator-runtime": "^0.12.1",
-    "sass": "^1.51.0",
-    "sass-lint": "^1.12.1",
+    "sass": "^1.69.1",
+    "sass-lint": "^1.13.1",
     "sass-lint-auto-fix": "^0.21.0",
-    "sass-loader": "^10.2.1",
+    "sass-loader": "^13.3.2",
     "source-map-loader": "^0.2.4",
-    "style-loader": "^0.22.1",
-    "svg-inline-loader": "^0.8.0",
+    "stream-browserify": "^3.0.0",
+    "style-loader": "^3.3.3",
+    "svg-inline-loader": "^0.8.2",
     "svg-inline-react": "^3.1.0",
-    "webpack": "^4.46.0",
-    "webpack-bundle-analyzer": "^3.3.2",
-    "webpack-cli": "^3.3.11",
-    "webpack-dev-server": "^3.11.0",
-    "yaml-loader": "^0.6.0"
-  },
-  "babel": {
-    "presets": [
-      "@babel/preset-env",
-      "@babel/preset-react",
-      "babel-preset-mobx"
-    ],
-    "plugins": [
-      "syntax-async-functions",
-      "@babel/plugin-syntax-dynamic-import",
-      "@babel/plugin-syntax-jsx",
-      "@babel/plugin-syntax-object-rest-spread",
-      "@babel/plugin-proposal-object-rest-spread",
-      "@babel/plugin-transform-react-jsx",
-      "@babel/plugin-transform-regenerator"
-    ]
+    "url": "^0.11.3",
+    "webpack": "^5.88.2",
+    "webpack-bundle-analyzer": "^4.9.1",
+    "webpack-cli": "^5.1.4",
+    "webpack-dev-server": "^4.15.1",
+    "yaml-loader": "^0.8.0"
   }
 }
