@@ -3,89 +3,74 @@ import "./static/stylesheets/defaults.scss";
 
 import React from "react";
 import { render } from "react-dom";
-import { Route, Switch, HashRouter } from "react-router-dom";
-import { Redirect } from "react-router";
+import {BrowserRouter} from "react-router-dom";
 
-import * as Stores from "./stores";
-import {inject, observer, Provider} from "mobx-react";
+import {observer} from "mobx-react";
+import {accountsStore, rootStore} from "./stores";
 
-import {Action, ErrorHandler, LoadingElement} from "elv-components-js";
+import {Action, ErrorHandler} from "elv-components-js";
 
-import EnforceLogin from "./EnforceLogin";
+import Header from "./Header";
+import Navigation from "./Navigation";
 
-import Header from "./components/Header";
-import Navigation from "./components/Navigation";
+import AppRoutes from "./Routes";
+import {Group, Loader, MantineProvider} from "@mantine/core";
+import ScrollToTop from "./ScrollToTop";
 
-import {AppRoutes, SiteRoutes} from "./Routes";
-
-@inject("rootStore")
-@inject("accountsStore")
-@observer
-class App extends React.PureComponent {
-  render() {
-    if(this.props.rootStore.configError) {
-      return (
-        <div className="page-error">
-          <div className="page-error-container">
-            <h1>
-              Unable to load client configuration
-            </h1>
-            <h1>
-              {EluvioConfiguration["config-url"]}
-            </h1>
-
-            <Action onClick={() => this.props.rootStore.InitializeClient()}>
-              Retry
-            </Action>
-          </div>
-        </div>
-      );
-    }
-
+const App = observer(() => {
+  if(rootStore.configError) {
     return (
-      <HashRouter>
-        <LoadingElement
-          loading={!this.props.rootStore.client || !this.props.accountsStore.accountsLoaded}
-          fullPage={true}
-          render={() => (
-            <div className="router-container">
-              <Header />
-              <Navigation />
-              <EnforceLogin>
-                <Switch>
-                  {
-                    SiteRoutes.map(({path, component}) =>
-                      <Route key={`route-${path}`} exact path={path} component={component}/>
-                    )
-                  }
+      <div className="page-error">
+        <div className="page-error-container">
+          <h1>
+            Unable to load client configuration
+          </h1>
+          <h1>
+            {EluvioConfiguration["config-url"]}
+          </h1>
 
-                  {
-                    AppRoutes.map(({path, component}) =>
-                      <Route key={`route-${path}`} exact path={path} component={component}/>
-                    )
-                  }
-
-                  <Route>
-                    <Redirect to="/accounts"/>
-                  </Route>
-                </Switch>
-              </EnforceLogin>
-            </div>
-          )}
-        />
-      </HashRouter>
+          <Action onClick={() => rootStore.InitializeClient()}>
+            Retry
+          </Action>
+        </div>
+      </div>
     );
   }
-}
+
+  if(!rootStore.client || !accountsStore.accountsLoaded) {
+    return (
+      <Group h="100vh" align="center" position="center">
+        <Loader />
+      </Group>
+    );
+  }
+
+  return (
+    <BrowserRouter>
+      <ScrollToTop />
+      <div className="router-container">
+        <Header />
+        <Navigation />
+        <AppRoutes />
+      </div>
+    </BrowserRouter>
+  );
+});
 
 const AppComponent = ErrorHandler(App);
+
+const element = document.createElement("div");
+element.id = "app";
+document.body.appendChild(
+  element
+);
 
 render(
   (
     <React.Fragment>
-      <Provider {...Stores}>
+      <MantineProvider withGlobalStyles>
         <AppComponent />
-      </Provider>
+      </MantineProvider>
       <div className="app-version">{EluvioConfiguration.version}</div>
     </React.Fragment>
   ),
