@@ -1,104 +1,129 @@
-const webpack = require("webpack");
 const Path = require("path");
-const autoprefixer = require("autoprefixer");
 const BundleAnalyzerPlugin = require("webpack-bundle-analyzer").BundleAnalyzerPlugin;
 const HtmlWebpackPlugin = require("html-webpack-plugin");
-const CopyWebpackPlugin = require("copy-webpack-plugin");
+const ReactRefreshWebpackPlugin = require("@pmmmwh/react-refresh-webpack-plugin");
 
-module.exports = {
-  entry: "./src/index.js",
-  target: "web",
-  output: {
-    path: Path.resolve(__dirname, "dist"),
-    filename: "index.js",
-    chunkFilename: "[name].bundle.js"
-  },
-  devServer: {
-    disableHostCheck: true,
-    inline: true,
-    port: 8080,
-    headers: {
-      "Access-Control-Allow-Origin": "*",
-      "Access-Control-Allow-Headers": "Content-Type, Accept",
-      "Access-Control-Allow-Methods": "POST"
-    }
-  },
-  node: {
-    fs: "empty"
-  },
-  mode: "development",
-  devtool: "eval-source-map",
-  optimization: {
-    splitChunks: {
-      chunks: "all"
-    }
-  },
-  plugins: [
-    new CopyWebpackPlugin([{
-      from: Path.join(__dirname, "configuration.js"),
-      to: Path.join(__dirname, "dist", "configuration.js")
-    }]),
+module.exports = env => {
+  const isDevelopment = !!env.WEBPACK_SERVE;
+
+  let plugins = [
     new HtmlWebpackPlugin({
-      title: "Eluvio",
+      title: "Eluvio Core",
       template: Path.join(__dirname, "src", "index.html"),
-      inject: "body",
-      cache: false,
       filename: "index.html",
-      favicon: Path.join(__dirname, "src", "static", "icons", "favicon-light.png")
-    }),
-  ],
-  module: {
-    rules: [
-      {
-        test: /\.scss$/,
-        use: [
-          "style-loader",
-          {
-            loader: "css-loader",
-            options: {
-              importLoaders: 2
-            }
-          },
-          {
-            loader: "postcss-loader",
-            options: {
-              plugins: () => [autoprefixer({})]
-            }
-          },
-          "sass-loader"
-        ]
-      },
-      {
-        test: /\.(js|mjs)$/,
-        exclude: /node_modules\/(?!elv-components-js)/,
-        loader: "babel-loader",
-        options: {
-          presets: ["@babel/preset-env", "@babel/preset-react", "babel-preset-mobx"],
-          plugins: [
-            require("@babel/plugin-proposal-object-rest-spread"),
-            require("@babel/plugin-transform-regenerator"),
-            require("@babel/plugin-transform-runtime"),
-            require("@babel/plugin-syntax-dynamic-import")
-          ]
-        }
-      },
-      {
-        test: /\.svg$/,
-        loader: "svg-inline-loader"
-      },
-      {
-        test: /\.(gif|png|jpe?g)$/i,
-        use: ["file-loader"],
-      },
-      {
-        test: /\.(txt|bin|abi|html)$/i,
-        loader: "raw-loader"
-      },
-      {
-        test: /\.ya?ml$/,
-        type: "json", // Required by Webpack v4
-        use: "yaml-loader"
-      }
-    ]
+      favicon: Path.join(__dirname, "src", "static", "icons", "favicon-light.png"),
+      inject: "body"
+    })
+  ];
+
+  if(process.env.ANALYZE_BUNDLE) {
+    plugins.push(new BundleAnalyzerPlugin());
   }
+
+  if(isDevelopment) {
+    plugins.push(new ReactRefreshWebpackPlugin());
+  }
+
+  return {
+    entry: Path.resolve(__dirname, "src/index.js"),
+    target: "web",
+    output: {
+      path: Path.resolve(__dirname, "dist"),
+      clean: true,
+      //filename: "main.js",
+      filename: "[name].bundle.js",
+      //publicPath: process.env.ASSET_PATH,
+      publicPath: "/",
+      chunkFilename: "bundle.[id].[chunkhash].js"
+    },
+    devServer: {
+      hot: true,
+      historyApiFallback: true,
+      allowedHosts: "all",
+      port: 8082,
+      headers: {
+        "Access-Control-Allow-Origin": "*",
+        "Access-Control-Allow-Headers": "Content-Type, Accept",
+        "Access-Control-Allow-Methods": "POST"
+      },
+      // This is to allow configuration.js to be accessed
+      static: {
+        directory: Path.resolve(__dirname, "./config"),
+        publicPath: "/"
+      }
+    },
+    mode: "development",
+    devtool: "eval-source-map",
+    plugins,
+    externals: {
+      crypto: "crypto"
+    },
+    optimization: {
+      splitChunks: {
+        chunks: "all",
+      },
+    },
+    resolve: {
+      fallback: {
+        stream: require.resolve("stream-browserify"),
+        url: require.resolve("url")
+      },
+      extensions: [".js", ".jsx", ".mjs", ".scss", ".png", ".svg"],
+    },
+    module: {
+      rules: [
+        {
+          test: /\.(css|scss)$/,
+          exclude: /\.(theme|font)\.(css|scss)$/i,
+          use: [
+            "style-loader",
+            {
+              loader: "css-loader",
+              options: {
+                importLoaders: 2
+              }
+            },
+            "postcss-loader",
+            "sass-loader"
+          ]
+        },
+        {
+          test: /\.(js|mjs|jsx)$/,
+          loader: "babel-loader",
+          options: {
+            plugins: [isDevelopment && require.resolve("react-refresh/babel")].filter(Boolean),
+            presets: [
+              "@babel/preset-env",
+              "@babel/preset-react",
+            ]
+          }
+        },
+        {
+          test: /\.svg$/,
+          loader: "svg-inline-loader"
+        },
+        {
+          test: /\.(gif|png|jpe?g|otf|woff2?|ttf)$/i,
+          include: [Path.resolve(__dirname, "src/static/public")],
+          type: "asset/inline",
+          generator: {
+            filename: "public/[name][ext]"
+          }
+        },
+        {
+          test: /\.(gif|png|jpe?g|otf|woff2?|ttf)$/i,
+          type: "asset/resource",
+        },
+        {
+          test: /\.(txt|bin|abi|html)$/i,
+          exclude: [Path.resolve(__dirname, "src/index.html")],
+          type: "asset/source"
+        },
+        {
+          test: /\.ya?ml$/,
+          use: "yaml-loader"
+        }
+      ]
+    }
+  };
 };
